working external entra id
